Struct icu_timezone::TimeZoneIdMapperWithFastCanonicalization

pub struct TimeZoneIdMapperWithFastCanonicalization<I> { /* private fields */ }

A mapper that supplements TimeZoneIdMapper with about 8 KB of additional data to improve the performance of canonical IANA ID lookup.

The data in TimeZoneIdMapper is optimized for IANA to BCP-47 lookup; the reverse requires a linear walk over all ~600 IANA identifiers. The data added here allows for constant-time mapping from BCP-47 to IANA.

Implementations

impl TimeZoneIdMapperWithFastCanonicalization<TimeZoneIdMapper>

pub fn new() -> Self

Creates a new TimeZoneIdMapperWithFastCanonicalization using compiled data.

See TimeZoneIdMapperWithFastCanonicalization for an example.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), ) -> Result<Self, TimeZoneError>

A version of [Self :: new] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), ) -> Result<Self, TimeZoneError>

A version of [Self :: new] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<P>(provider: &P) -> Result<Self, TimeZoneError>
where P: DataProvider<IanaToBcp47MapV2Marker> + DataProvider<Bcp47ToIanaMapV1Marker> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

impl<I> TimeZoneIdMapperWithFastCanonicalization<I>
where I: AsRef<TimeZoneIdMapper>,

pub fn try_new_with_mapper(mapper: I) -> Result<Self, TimeZoneError>

Creates a new TimeZoneIdMapperWithFastCanonicalization using compiled data and a pre-existing TimeZoneIdMapper, which can be borrowed.

See TimeZoneIdMapperWithFastCanonicalization for an example.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_mapper_with_any_provider( provider: &(impl AnyProvider + ?Sized), mapper: I, ) -> Result<Self, TimeZoneError>

A version of [Self :: try_new_with_mapper] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_mapper_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), mapper: I, ) -> Result<Self, TimeZoneError>

A version of [Self :: try_new_with_mapper] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_with_mapper_unstable<P>( provider: &P, mapper: I, ) -> Result<Self, TimeZoneError>
where P: DataProvider<IanaToBcp47MapV2Marker> + DataProvider<Bcp47ToIanaMapV1Marker> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn inner(&self) -> &TimeZoneIdMapper

Gets the inner TimeZoneIdMapper for performing queries.

pub fn as_borrowed( &self, ) -> TimeZoneIdMapperWithFastCanonicalizationBorrowed<'_>

Returns a borrowed version of the mapper that can be queried.

This avoids a small potential indirection cost when querying the mapper.